manana-seguro-bot/requirements.txt
python-telegram-bot==21.6
anthropic>=0.25.0
python-dotenv==1.0.0
stellar-sdk>=10.0.0
qrcode[pil]>=7.4.2
pillow>=10.0.0